Feature/applying ticket changes v6

See merge request veolia.com/vna/muni-water/brm-contract-ops/speridian/veolia-frontend!28

+++ src/app/clients/clients.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { SiteService } from '../sites/site.service';
 import { UserService } from '../users/user.service';
 import { ClientService } from './client.service';
 
@@ -11,7 +12,7 @@ type ClientType = 'true' | 'false';
   styleUrls: ['./clients.component.scss'],
 })
 export class ClientsComponent implements OnInit {
-  constructor(private fb: FormBuilder, private clientService: ClientService, private userService: UserService) {}
+  constructor(private fb: FormBuilder, private clientService: ClientService, private userService: UserService, private siteService: SiteService) {}
   form!: FormGroup;
   clientsArray: any[] = [];
   title: string = 'Clients';
@@ -54,12 +55,12 @@ export class ClientsComponent implements OnInit {
 
     this.clientService.postClient(this.form.value).subscribe({
       next: (_) => {
-        this.userService.openSnackBar('Client Created', 'close');
+        this.userService.openSnackBar('New Client is Created Successfully!', 'close');
         this.getClient();
       },
       error: (err: any) => {
-        window.location.reload();
-        this.error = err;
+        this.error = err.message;
+        this.userService.openSnackBar(this.error, 'close');
       },
     });
 
@@ -93,7 +94,7 @@ export class ClientsComponent implements OnInit {
       this.clientService.updateClient(this.currentClient, this.form.value)
         .subscribe({
           next: (_) => {
-            this.userService.openSnackBar('Client Updated', 'close');
+            this.userService.openSnackBar('Client is Updated Successfully!', 'close');
             this.getClient();
           },
           error: (err) => {
@@ -105,9 +106,36 @@ export class ClientsComponent implements OnInit {
     }
   }
 
-  onDeleteClient(id: any) {
-    this.clients = this.clients.filter(({ clientId }) => clientId != id);
 
-    this.clientService.deleteClient(id);
+  onDeleteClient(id: any) {
+
+    this.siteService.getSiteByClientId(id).subscribe(
+      (res:any)=>{
+        let sitesCount =  res.length
+        if (sitesCount>0) {
+          this.userService.openSnackBar('The client cannot be deleted until all the associated Sites are deleted or detached from the Client.', 'close');
+        }
+        else{
+
+          this.clients = this.clients.filter(({ clientId }) => clientId != id);
+          this.clientService.deleteClient(id);
+          this.userService.openSnackBar('Client Record Deleted Successfully!', 'close');
+
+        }
+      }
+    )
   }
+
+  // onDeleteClient(id: any) {
+
+  //   // if () {
+      
+  //   // }
+
+
+  //   this.clients = this.clients.filter(({ clientId }) => clientId != id);
+  //   this.userService.openSnackBar('Client Record is Deleted.', 'close');
+
+  //   this.clientService.deleteClient(id);
+  // }
 }
